services/observer/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
jinja2==3.1.3
httpx==0.27.0
pyyaml==6.0.1
